Move notifications utility into script.utils
